# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Modern Agent HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/sessions/{sessionId}/reviews:
    get:
      operationId: listReviews
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReviewsResponse'
          description: OK
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List a worker's code-review runs
      tags:
      - reviews
  /api/v1/sessions/{sessionId}/reviews/submit:
    post:
      operationId: submitReview
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReviewInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewRunResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Record a reviewer's result for a worker's PR
      tags:
      - reviews
  /api/v1/sessions/{sessionId}/reviews/trigger:
    post:
      operationId: triggerReview
      parameters:
      - description: Session identifier, e.g. project-1.
        in: path
        name: sessionId
        required: true
        schema:
          description: Session identifier, e.g. project-1.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerReviewResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerReviewResponse'
          description: Created
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Trigger a code review of a worker's PR
      tags:
      - reviews
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    ListReviewsResponse:
      properties:
        reviewerHandleId:
          type: string
        reviews:
          items:
            $ref: '#/components/schemas/PRReviewState'
          type: array
      required:
      - reviewerHandleId
      - reviews
      type: object
...
    PRReviewState:
      properties:
        latestRun:
          $ref: '#/components/schemas/ReviewRun'
        prNumber:
          type: integer
        prUrl:
          type: string
        status:
          enum:
          - needs_review
          - running
          - up_to_date
          - changes_requested
          - ineligible
          type: string
        targetSha:
          type: string
        title:
          type: string
      required:
      - prUrl
      - prNumber
      - title
      - targetSha
      - status
      type: object
...
    ReviewRun:
      properties:
        batchId:
          type: string
        body:
          type: string
        createdAt:
          format: date-time
          type: string
        deliveredAt:
          format: date-time
          type:
          - "null"
          - string
        githubReviewId:
          type: string
        harness:
          type: string
        id:
          type: string
        prUrl:
          type: string
        reviewId:
          type: string
        sessionId:
          type: string
        status:
          type: string
        targetSha:
          type: string
        verdict:
          type: string
      required:
      - id
      - reviewId
      - sessionId
      - batchId
      - harness
      - prUrl
      - targetSha
      - status
      - verdict
      - body
      - githubReviewId
      - createdAt
      type: object
    ReviewRunResponse:
      properties:
        review:
          $ref: '#/components/schemas/ReviewRun'
        reviewerHandleId:
          type: string
        reviews:
          items:
            $ref: '#/components/schemas/ReviewRun'
          type: array
      required:
      - review
      - reviews
      - reviewerHandleId
      type: object
...
    SubmitReviewInput:
      properties:
        body:
          description: Review body recorded by AO. Required for changes_requested.
          type: string
        githubReviewId:
          description: Id of the GitHub PR review the reviewer posted, if any.
          type: string
        reviews:
          description: Batched review results recorded by one reviewer CLI command.
          items:
            $ref: '#/components/schemas/SubmitReviewItem'
          type: array
        runId:
          description: Review run id being completed.
          type: string
        verdict:
          description: 'Review verdict: approved or changes_requested.'
          type: string
      type: object
    SubmitReviewItem:
      properties:
        body:
          description: Review body recorded by AO. Required for changes_requested.
          type: string
        githubReviewId:
          description: Id of the GitHub PR review the reviewer posted, if any.
          type: string
        runId:
          description: Review run id being completed.
          type: string
        verdict:
          description: 'Review verdict: approved or changes_requested.'
          type: string
      required:
      - runId
      - verdict
      type: object
...
    TriggerReviewResponse:
      properties:
        reviewerHandleId:
          type: string
        reviews:
          items:
            $ref: '#/components/schemas/PRReviewState'
          type: array
      required:
      - reviewerHandleId
      - reviews
      type: object
...